+++ crates/node/src/chain_spec.rs
@@ -58,6 +58,7 @@ pub fn authority_keys_from_seed(s: &str) -> (AuraId, GrandpaId) {
 pub fn development_config(sealing: SealingMode) -> Result<DevChainSpec, String> {
     let wasm_binary = WASM_BINARY.ok_or_else(|| "Development wasm not available".to_string())?;
     let chain_id = DEV_CHAIN_ID;
+    let genesis_loader = load_genesis()?;
 
     Ok(DevChainSpec::from_genesis(
         // Name
@@ -68,6 +69,7 @@ pub fn development_config(sealing: SealingMode) -> Result<DevChainSpec, String> 
         move || {
             DevGenesisExt {
                 genesis_config: testnet_genesis(
+                    genesis_loader.clone(),
                     wasm_binary,
                     // Initial PoA authorities
                     vec![authority_keys_from_seed("Alice")],
@@ -92,6 +94,7 @@ pub fn development_config(sealing: SealingMode) -> Result<DevChainSpec, String> 
 
 pub fn deoxys_config(sealing: SealingMode, chain_id: &str) -> Result<DevChainSpec, String> {
     let wasm_binary = WASM_BINARY.ok_or_else(|| "Development wasm not available".to_string())?;
+    let genesis_loader = load_genesis()?;
 
     Ok(DevChainSpec::from_genesis(
         // Name
@@ -102,6 +105,7 @@ pub fn deoxys_config(sealing: SealingMode, chain_id: &str) -> Result<DevChainSpe
         move || {
             DevGenesisExt {
                 genesis_config: testnet_genesis(
+                    genesis_loader.clone(),
                     wasm_binary,
                     // Initial PoA authorities
                     vec![authority_keys_from_seed("Alice")],
@@ -142,12 +146,12 @@ fn load_genesis() -> Result<GenesisData, String> {
 
 /// Configure initial storage state for FRAME modules.
 fn testnet_genesis(
+    genesis_loader: GenesisData,
     wasm_binary: &[u8],
     initial_authorities: Vec<(AuraId, GrandpaId)>,
     _enable_println: bool,
 ) -> RuntimeGenesisConfig {
-    // TODO: ensure this defaulted config isnt disturbing the genesis state
-    let starknet_genesis_config = GenesisConfig::default();
+    let starknet_genesis_config = GenesisConfig::from(genesis_loader);
 
     RuntimeGenesisConfig {
         system: SystemConfig {