+++ CHANGELOG.md
@@ -3,6 +3,7 @@ git # Deoxys Changelog
 ## Next release
 
 - refactor: optimise get_events RPC
+- fix(root): fixed state commitments broken due to genesis loader
 - feat(docker): add dockerfile and docker-compose
 - fix: fix implementation `get_storage_at()` for `BlockifierStateAdapter`
 - fix(sync): Fix end condition of the l2 sync